feat: add seed parameter to customize the rng (#39)
We would like to allow reproducibility of a result, for that, we need to
seed the RNG to a known state. Here we're adding a new parameter to the
different programs to allow seeding the RNG. On the other hand, as a
library, users can choose to seed the RNG or build it in the way they
want.